+++ src/TuneReporting/Base/Endpoints/AdvertiserReportActualsBase.php
@@ -2,7 +2,7 @@
 /**
  * AdvertiserReportActualsBase.php
  *
- * Copyright (c) 2014 TUNE, Inc.
+ * Copyright (c) 2015 TUNE, Inc.
  * All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -27,10 +27,10 @@
  *
  * @category  TUNE_Reporting
  * @author    Jeff Tanner <jefft@tune.com>
- * @copyright 2014 TUNE, Inc. (http://www.tune.com)
+ * @copyright 2015 TUNE, Inc. (http://www.tune.com)
  * @package   tune_reporting_base_endpoints
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   $Date: 2014-12-18 04:47:37 $
+ * @version   $Date: 2015-01-05 14:24:08 $
  * @link      https://developers.mobileapptracking.com/tune-reporting-sdks @endlink
  *
  */
@@ -287,7 +287,8 @@ abstract class AdvertiserReportActualsBase extends AdvertiserReportBase
         $client = new TuneManagementClient(
             "export",
             "download",
-            $this->api_key,
+            $this->auth_key,
+            $this->auth_type,
             $query_string_dict = array (
                 'job_id' => $job_id
             )