+++ tests/TestAdvertiserReportActuals.php
@@ -2,7 +2,7 @@
 /**
  * TestAdvertiserReportActuals.php, TUNE Reporting SDK PHPUnit Test
  *
- * Copyright (c) 2014 TUNE, Inc.
+ * Copyright (c) 2015 TUNE, Inc.
  * All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -28,9 +28,9 @@
  * @category  TUNE_Reporting
  *
  * @author    Jeff Tanner <jefft@tune.com>
- * @copyright 2014 TUNE, Inc. (http://www.tune.com)
+ * @copyright 2015 TUNE, Inc. (http://www.tune.com)
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   $Date: 2015-01-03 08:14:06 $
+ * @version   $Date: 2015-01-05 14:24:08 $
  * @link      https://developers.mobileapptracking.com/tune-reporting-sdks @endlink
  *
  */
@@ -38,6 +38,7 @@
 require_once dirname(__FILE__) . "/../src/TuneReporting.php";
 
 use TuneReporting\Api\AdvertiserReportActuals;
+use TuneReporting\Api\SessionAuthenticate;
 use TuneReporting\Helpers\SdkConfig;
 
 class TestAdvertiserReportActuals extends \PHPUnit_Framework_TestCase
@@ -46,6 +47,16 @@ class TestAdvertiserReportActuals extends \PHPUnit_Framework_TestCase
      * @ignore
      */
     protected $advertiser_report = null;
+
+    /**
+     * @ignore
+     */
+    protected $api_key = null;
+
+    /**
+     * @ignore
+     */
+    protected $session_token = null;
 
     /**
      * Get API Key from environment.
@@ -59,6 +70,15 @@ class TestAdvertiserReportActuals extends \PHPUnit_Framework_TestCase
 
         $api_key = getenv('API_KEY');
         $this->assertNotNull($api_key);
+        $this->api_key = $api_key;
+
+        $session_authenticate = new SessionAuthenticate();
+        $response = $session_authenticate->api_key($api_key);
+        $this->assertNotNull($response);
+        $session_token = $response->getData();
+        $this->assertNotNull($session_token);
+        $this->session_token = $session_token;
+
         $tune_reporting_config_file = dirname(__FILE__) . "/../config/tune_reporting_sdk.config";
         $this->assertTrue(file_exists($tune_reporting_config_file), "SDK config file does not exist: '{$tune_reporting_config_file}'");
         $sdk_config = SdkConfig::getInstance($tune_reporting_config_file);
@@ -76,12 +96,14 @@ class TestAdvertiserReportActuals extends \PHPUnit_Framework_TestCase
     {
         $sdk_config = $this->advertiser_report->getSdkConfig();
         $this->assertNotNull($sdk_config);
-        $api_key = $sdk_config->getApiKey();
+        $auth_key = $sdk_config->getAuthKey();
+        $auth_type = $sdk_config->getAuthType();
 
-        $this->assertNotNull($api_key, "In tune_reporting_sdk.config, set 'tune_reporting_api_key_string'");
-        $this->assertInternalType('string', $api_key, "In tune_reporting_sdk.config, set 'tune_reporting_api_key_string'");
-        $this->assertNotEmpty($api_key, "In tune_reporting_sdk.config, set 'tune_reporting_api_key_string'");
-        $this->assertNotEquals("API_KEY", $api_key, "In tune_reporting_sdk.config, set 'tune_reporting_api_key_string'");
+        $this->assertNotNull($auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertInternalType('string', $auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertNotEmpty($auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertNotEquals("API_KEY", $auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertEquals("api_key", $auth_type, "In tune_reporting_sdk.config, set 'tune_reporting_auth_type_string'");
     }
 
     /**
@@ -155,6 +177,7 @@ class TestAdvertiserReportActuals extends \PHPUnit_Framework_TestCase
         $this->assertNotNull($response);
         $this->assertEquals(200, $response->getHttpCode());
     }
+
 
     /**
      * Test find
@@ -440,5 +463,75 @@ class TestAdvertiserReportActuals extends \PHPUnit_Framework_TestCase
         } catch (Exception $ex ) {
             $this->fail($ex->getMessage());
         }
+    }
+
+    /**
+     * Test session token
+     */
+    public function testSessionToken()
+    {
+        $sdk_config = SdkConfig::getInstance();
+        $this->assertNotNull($sdk_config);
+        $sdk_config->setSessionToken($this->session_token);
+
+        $auth_key = $sdk_config->getAuthKey();
+        $auth_type = $sdk_config->getAuthType();
+
+        $this->assertNotNull($auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertInternalType('string', $auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertNotEmpty($auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertNotEquals("API_KEY", $auth_key, "In tune_reporting_sdk.config, set 'tune_reporting_auth_key_string'");
+        $this->assertEquals("session_token", $auth_type, "In tune_reporting_sdk.config, set 'tune_reporting_auth_type_string'");
+    }
+
+
+    /**
+     * Test count
+     */
+    public function testCountSessionToken()
+    {
+        $week_ago       = date('Y-m-d', strtotime("-8 days"));
+        $yesterday      = date('Y-m-d', strtotime("-1 days"));
+        $start_date     = "{$week_ago} 00:00:00";
+        $end_date       = "{$yesterday} 23:59:59";
+
+        $response = $this->advertiser_report->count(
+            $start_date,
+            $end_date,
+            $group               = "site_id,publisher_id",
+            $filter              = "(publisher_id > 0)",
+            $response_timezone   = "America/Los_Angeles"
+        );
+
+        $this->assertNotNull($response);
+        $this->assertEquals(200, $response->getHttpCode());
+    }
+
+
+    /**
+     * Test find
+     */
+    public function testFindSessionToken()
+    {
+        $week_ago       = date('Y-m-d', strtotime("-8 days"));
+        $yesterday      = date('Y-m-d', strtotime("-1 days"));
+        $start_date     = "{$week_ago} 00:00:00";
+        $end_date       = "{$yesterday} 23:59:59";
+
+        $response = $this->advertiser_report->find(
+            $start_date,
+            $end_date,
+            $fields              = null,
+            $group               = "site_id,publisher_id",
+            $filter              = "(publisher_id > 0)",
+            $limit               = 5,
+            $page                = null,
+            $sort                = array("installs" => "DESC"),
+            $timestamp           = "datehour",
+            $response_timezone   = "America/Los_Angeles"
+        );
+
+        $this->assertNotNull($response);
+        $this->assertEquals(200, $response->getHttpCode());
     }
 }